backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
stripe
cloudinary
stripe==12.5.1
